requirements.txt
altgraph==0.17.3
MouseInfo==0.1.3
paho-mqtt==1.6.1
pefile==2023.2.7
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
pytweening==1.0.4
pywin32-ctypes==0.2.0
